Ex09.c

+++ Ex09.c
@@ -33,7 +33,7 @@ float imposto(float sal1){
    }  
 }
 
-float main(){
+int main(){
   float sal;
 
     printf("Informe o salário: ");
